Move Dockerfile copy/add --exclude implementation to Labs
