+++ client/llb/fileop.go
@@ -398,16 +398,16 @@ func WithAllowWildcard(b bool) RmOption {
 	})
 }
 
-type excludeOnCopy struct {
+type excludeOnCopyAction struct {
 	patterns []string
 }
 
-func (e *excludeOnCopy) SetCopyOption(i *CopyInfo) {
+func (e *excludeOnCopyAction) SetCopyOption(i *CopyInfo) {
 	i.ExcludePatterns = append(i.ExcludePatterns, e.patterns...)
 }
 
 func WithExcludePatterns(patterns []string) CopyOption {
-	return &excludeOnCopy{patterns}
+	return &excludeOnCopyAction{patterns}
 }
 
 type fileActionRm struct {
