+++ frontend/dockerfile/dockerfile2llb/convert_test.go
@@ -208,19 +208,3 @@ COPY --from=stage1 f2 /sub/
 	_, _, _, err = Dockerfile2LLB(appcontext.Context(), []byte(df), ConvertOpt{})
 	assert.EqualError(t, err, "circular dependency detected on stage: stage0")
 }
-
-func TestDockerfileCopyExcludePatterns(t *testing.T) {
-	df := `FROM scratch
-COPY --exclude=src/*.go --exclude=tmp/*.txt dir /sub/
-`
-	_, _, _, err := Dockerfile2LLB(appcontext.Context(), []byte(df), ConvertOpt{})
-	assert.NoError(t, err)
-}
-
-func TestDockerfileAddExcludePatterns(t *testing.T) {
-	df := `FROM scratch
-ADD --exclude=src/*.go --exclude=tmp/*.txt dir /sub/
-`
-	_, _, _, err := Dockerfile2LLB(appcontext.Context(), []byte(df), ConvertOpt{})
-	assert.NoError(t, err)
-}
