+++ frontend/dockerfile/instructions/exclude_pattern_feature.go
@@ -1,0 +1,8 @@
+//go:build dfexcludepatterns
+// +build dfexcludepatterns
+
+package instructions
+
+func init() {
+	excludePatternsEnabled = true
+}
